LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-copyTo.crossOriginIsolated.https.any.worker-expected.txt


FAIL Test copying I420 frame to SharedArrayBuffer. assert_equals: allocationSize() expected 12 but got 0
FAIL Test copying I420 frame to shared ArrayBufferView. assert_equals: allocationSize() expected 12 but got 0
